fix: floatx II=1, optimizing linalg (WIP)

+++ tests/hls/linalg/determinant/top.cpp
@@ -7,7 +7,8 @@ extern "C" void top(const double in_a[16],
 #pragma HLS INTERFACE s_axilite port = out bundle = control
 #pragma HLS INTERFACE s_axilite port = return bundle = control
 
-    // determinant
-    linalg::Mat3<double> M(in_a);
+//#pragma HLS PIPELINE
+
+    linalg::Mat3<float> M(in_a);
     out = (double)M.determinant();
 }
